Moves LLMMessage to avoid a cicular import with LLM classes
